third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/stylenonce-basic-blocked-error-event.html

<!DOCTYPE html>
<html>
<head>
    <title>Style element has error on bad style nonce</title>
    <meta http-equiv="Content-Security-Policy" content="style-src 'nonce-nonceynonce'">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script>
        function styleError() {
            var color = window.getComputedStyle(document.querySelector('p')).color;
            assert_equals(color, "rgb(0, 128, 0)", "The color of the paragraph is still green.");
            done();
        }
    </script>
    <style nonce="nonceynonce">p { color: green; }</style>
    <style nonce="notavalidnonce" onerror="styleError();">p { color: red; }</style>
</head>
    <p>A test paragraph</p>
</html>
